refactor(web): encapsulates non-embedded keytip
